RIB-91: refactor PlaceService
